Redid android deployment for newer Qt versions
The new deployment is a lot smarter and does a lot of stuff on its
own...

The easiest way to make use of it was to simply re-setup the config
using qt-creator.

The android configuration is now located under example/ (like tizen).

Also added higher resolution icons and removed lower resolution one (as
suggested by the android icon style document).

Fixes #4.
Fixes #6.