Fixed wrapping of long lines
Since there is no width set the Text item simply expanded over the
bounds of the actually usable area. However setting the width is
destroying the effect the anchor.centerIn had, since it centered the
label (which was left aligned) in the button (result in a normal left
alignen label basically).

So I had to specify horizontalAlignment too in order to get the
additional lines of the label centered.

the anchors.centerIn however still cannot be dropped since the label
would then be a few pixels too low.